Move SSPPick() class definition to its own file
Previously, the class was called Pick()
